feat: add output and dryrun options to batch delete

+++ src/Api/Batch.php
@@ -9,6 +9,13 @@ class Batch extends Endpoint
 {
     public const ENDPOINT = 'batch/objects';
 
+    protected const OUTPUT_METHODS = [
+        'minimal',
+        'verbose',
+    ];
+
+    protected const DEFAULT_OUTPUT = 'minimal';
+
     public function create(array $objects): ObjectCollection
     {
         return new ObjectCollection(
@@ -18,13 +25,23 @@ class Batch extends Endpoint
         );
     }
 
-    public function delete(string $className, array $where = []): Response
+    public function delete(string $className, array $where = [], ?string $output = self::DEFAULT_OUTPUT, bool $dryRun = false): Response
     {
+        if (is_null($output)) {
+            $output = self::DEFAULT_OUTPUT;
+        }
+
+        if (!in_array($output, self::OUTPUT_METHODS)) {
+            throw new \InvalidArgumentException('Invalid output method');
+        }
+
         return $this->api->delete(self::ENDPOINT, [
             'match' => [
                 'class' => $className,
                 'where' => $where,
             ],
+            'output' => $output,
+            'dryRun' => $dryRun,
         ]);
     }
 }
